Further commonization!
toroplus was previously using that weird, large gps.omap4.so without the lib_gsd4t.so. Turns out, the Sprint stock toroplus ROM uses the small gps.omap4.so plus large lib_gsd4t.so combo. The GPS biniares are not included in the 'Binaries for Nexus Devices' page, and considering the large gps.omap4.so (which maguro uses as well, IN THE STOCK MAGURO ROM) was working fine on toroplus for all this time, I can only conclude that Google came out with a better GPS setup after the launch of maguro, which previously DID use the 2 binary approach. Therefore I am switching all tuna devices to it.

Also of note, the toroplus stock ROM doesn't even include libdrmdecrypt.so in the first place, so it can just be commonized as maguro and toro use the same one.